# Conflicts:
#	app/models/user/constants.rb

+++ app/models/user.rb
@@ -140,7 +140,7 @@ class User < ActiveRecord::Base
 
   attr_accessor :import, :highlight_name, :highlight_job_title, :created_from_email, :sbrr_fresh_user,
                 :primary_email_attributes, :tags_updated, :keep_user_active, :escape_liquid_attributes, 
-                :role_ids_changed, :detect_language, :tag_use_updated 
+                :role_ids_changed, :detect_language, :tag_use_updated, :user_companies_updated 
                 # (This role_ids_changed used to forcefully call user callbacks only when role_ids are there.
   # As role_ids are not part of user_model(it is an association_reader), 
   #   agent.update_attributes won't trigger user callbacks since user doesn't have any change.
@@ -1111,7 +1111,7 @@ class User < ActiveRecord::Base
   end
 
   def create_freshid_user
-    return unless freshid_enabled_and_agent?
+    return if freshid_disabled_or_customer?
     Rails.logger.info "FRESHID Creating user :: a=#{self.account_id}, u=#{self.id}, email=#{self.email}"
     reset_freshid_user if email_changed?
     freshid_user = Freshid::User.create(user_attributes_for_freshid)
@@ -1129,7 +1129,7 @@ class User < ActiveRecord::Base
   end
 
   def destroy_freshid_user
-    if freshid_enabled_account? && freshid_authorization.present?
+    if freshid_enabled_account? && email_allowed_in_freshid? && freshid_authorization.present?
       remove_freshid_user
       freshid_authorization.destroy
       self.password_salt = self.crypted_password = nil
@@ -1178,11 +1178,15 @@ class User < ActiveRecord::Base
     end
 
     def freshid_enabled_and_agent?
-      agent? && freshid_enabled_account?
+      agent? && freshid_enabled_account? && email_allowed_in_freshid?
     end
 
-    def freshid_disabled_and_customer?
+    def freshid_disabled_or_customer?
       !freshid_enabled_and_agent?
+    end
+
+    def email_allowed_in_freshid?
+      !FRESHID_IGNORED_EMAIL_IDS.include?(self.email)
     end
 
     def valid_freshid_login?(incoming_password)
